+++ R/vis_spatial_gg.R
@@ -292,7 +292,7 @@
                 yend = sdimy_end
             ),
             color = network_color,
-            size = 0.5,
+            linewidth = 0.5,
             alpha = network_alpha
         )
     }
@@ -868,7 +868,7 @@ spatPlot <- function(...) {
 #' @param sdimx x-axis dimension name (default = 'sdimx')
 #' @param sdimy y-axis dimension name (default = 'sdimy')
 #' @param line_color color of line within pie charts
-#' @param radius radios of pie charts
+#' @param pie_scale amount to scale the pie size if there is no radius mapping exists
 #' @param alpha alpha of pie charts
 #' @param coord_fix_ratio fix ratio between x and y-axis
 #' @param title title of plot
@@ -894,7 +894,7 @@ spatDeconvPlot <- function(
         sdimy = "sdimy",
         cell_color_code = NULL,
         line_color = NA,
-        radius = 10,
+        pie_scale = 0.35,
         alpha = 1,
         legend_text = 8,
         background_color = "white",
@@ -963,9 +963,6 @@ spatDeconvPlot <- function(
         output = "data.table"
     )
 
-
-
-
     ### create 2D plot with ggplot ###
 
     pl <- ggplot2::ggplot()
@@ -992,7 +989,7 @@ spatDeconvPlot <- function(
         sdimy = sdimy,
         spatial_locations = spatial_locations,
         spatial_enrichment = spatial_enrichment,
-        radius = radius,
+        pie_scale = pie_scale,
         color = line_color,
         alpha = alpha,
         cell_color_code = cell_color_code
@@ -1982,7 +1979,7 @@ plotPCA <- function(
 #' @param sdimy = spatial dimension to use on y-axis
 #' @param spat_point_shape shape of points (border, no_border or voronoi)
 #' @param spat_point_size size of spatial points
-#' @param spat_point_alpha transparancy of spatial points
+#' @param spat_point_alpha transparency of spatial points
 #' @param spat_point_border_col border color of spatial points
 #' @param spat_point_border_stroke border stroke of spatial points
 #' @param dim_show_cluster_center show the center of each cluster
@@ -2637,14 +2634,14 @@ spatFeatPlot2D_single <- function(
                 yend <- paste0(sdimy, "_end")
                 pl <- pl + ggplot2::geom_segment(
                     data = spatial_network,
-                    aes_string(
+                    aes_string2(
                         x = xbegin,
                         y = ybegin,
                         xend = xend,
                         yend = yend
                     ),
                     color = network_color,
-                    size = 0.5,
+                    linewidth = 0.5,
                     alpha = edge_alpha
                 )
             }
@@ -2672,7 +2669,7 @@ spatFeatPlot2D_single <- function(
 
                 pl <- pl + ggplot2::geom_rect(
                     data = spatial_grid,
-                    aes_string(
+                    aes_string2(
                         xmin = xmin,
                         xmax = xmax,
                         ymin = ymin,
@@ -2772,9 +2769,10 @@ spatFeatPlot2D_single <- function(
             if (scale_alpha_with_expression == TRUE) {
                 pl <- pl + ggforce::geom_voronoi_tile(
                     data = cell_locations_metadata_feats,
-                    aes_string(
-                        x = sdimx, y = sdimy,
-                        group = "-1L",
+                   aes_string2(
+                        x = sdimx,
+                        y = sdimy,
+                        group = -1L,
                         fill = feat,
                         alpha = feat
                     ),
@@ -2785,9 +2783,10 @@ spatFeatPlot2D_single <- function(
             } else {
                 pl <- pl + ggforce::geom_voronoi_tile(
                     data = cell_locations_metadata_feats,
-                    aes_string(
-                        x = sdimx, y = sdimy,
-                        group = "-1L",
+                    aes_string2(
+                        x = sdimx,
+                        y = sdimy,
+                        group = -1L,
                         fill = feat
                     ),
                     colour = vor_border_color,
@@ -2808,11 +2807,16 @@ spatFeatPlot2D_single <- function(
                 xend <- paste0(sdimx, "_end")
                 yend <- paste0(sdimy, "_end")
                 pl <- pl + ggplot2::geom_segment(
-                    data = spatial_network, aes_string(
-                        x = xbegin, y = ybegin,
-                        xend = xend, yend = yend
+                    data = spatial_network,
+                    aes_string2(
+                        x = xbegin,
+                        y = ybegin,
+                        xend = xend,
+                        yend = yend
                     ),
-                    color = network_color, size = 0.5, alpha = 0.5
+                    color = network_color,
+                    linewidth = 0.5,
+                    alpha = 0.5
                 )
             }
 
@@ -2838,11 +2842,13 @@ spatFeatPlot2D_single <- function(
                 ), "_end")
 
                 pl <- pl + ggplot2::geom_rect(
-                    data = spatial_grid, aes_string(
+                    data = spatial_grid,
+                    aes_string2(
                         xmin = xmin, xmax = xmax,
                         ymin = ymin, ymax = ymax
                     ),
-                    color = grid_color, fill = NA
+                    color = grid_color,
+                    fill = NA
                 )
             }
 
@@ -3252,31 +3258,37 @@ spatFeatPlot2D <- function(
             edge_alpha <- 0.5
             pl <- pl + ggplot2::geom_segment(
                 data = annotated_network_DT,
-                aes_string(
+                aes_string2(
                     x = from_dim_names[1], y = from_dim_names[2],
                     xend = to_dim_names[1], yend = to_dim_names[2]
                 ),
-                alpha = edge_alpha, color = network_color, size = 0.1,
+                alpha = edge_alpha,
+                color = network_color,
+                linewidth = 0.1,
                 show.legend = FALSE
             )
         } else if (is.numeric(edge_alpha)) {
             pl <- pl + ggplot2::geom_segment(
                 data = annotated_network_DT,
-                aes_string(
+                aes_string2(
                     x = from_dim_names[1], y = from_dim_names[2],
                     xend = to_dim_names[1], yend = to_dim_names[2]
                 ),
-                alpha = edge_alpha, color = network_color, size = 0.1,
+                alpha = edge_alpha,
+                color = network_color,
+                linewidth = 0.1,
                 show.legend = FALSE
             )
         } else if (is.character(edge_alpha)) {
             if (edge_alpha %in% colnames(annotated_network_DT)) {
                 pl <- pl + ggplot2::geom_segment(
                     data = annotated_network_DT,
-                    aes_string(
-                        x = from_dim_names[1], y = from_dim_names[2],
+                    aes_string2(
+                        x = from_dim_names[1],
+                        y = from_dim_names[2],
                         xend = to_dim_names[1],
-                        yend = to_dim_names[2], alpha = edge_alpha
+                        yend = to_dim_names[2],
+                        alpha = edge_alpha
                     ),
                     color = network_color,
                     show.legend = FALSE
@@ -3292,9 +3304,11 @@ spatFeatPlot2D <- function(
         message("no feats selected")
         pl <- pl + ggplot2::geom_point(
             data = data,
-            aes_string(x = dim_names[1], dim_names[2]),
-            fill = cell_color, show.legend = show_legend,
-            size = point_size, alpha = point_alpha
+            aes_string2(x = dim_names[1], dim_names[2]),
+            fill = cell_color,
+            show.legend = show_legend,
+            size = point_size,
+            alpha = point_alpha
         )
     } else {
         ## set gradient limits if needed ##
@@ -3321,25 +3335,32 @@ spatFeatPlot2D <- function(
         if (point_shape == "border") {
             if (scale_alpha_with_expression == TRUE) {
                 pl <- pl + ggplot2::geom_point(
-                    data = data, aes_string2(
+                    data = data,
+                    aes_string2(
                         x = dim_names[1],
                         y = dim_names[2],
-                        fill = feat, alpha = feat
+                        fill = feat,
+                        alpha = feat
                     ),
-                    show.legend = show_legend, shape = 21,
+                    show.legend = show_legend,
+                    shape = 21,
                     size = point_size,
-                    color = point_border_col, stroke = point_border_stroke
+                    color = point_border_col,
+                    stroke = point_border_stroke
                 )
             } else {
                 pl <- pl + ggplot2::geom_point(
-                    data = data, aes_string2(
+                    data = data,
+                    aes_string2(
                         x = dim_names[1],
                         y = dim_names[2],
                         fill = feat
                     ),
-                    show.legend = show_legend, shape = 21,
+                    show.legend = show_legend,
+                    shape = 21,
                     size = point_size,
-                    color = point_border_col, stroke = point_border_stroke,
+                    color = point_border_col,
+                    stroke = point_border_stroke,
                     alpha = point_alpha
                 )
             }
@@ -3361,21 +3382,27 @@ spatFeatPlot2D <- function(
         if (point_shape == "no_border") {
             if (scale_alpha_with_expression == TRUE) {
                 pl <- pl + ggplot2::geom_point(
-                    data = data, aes_string2(
+                    data = data,
+                    aes_string2(
                         x = dim_names[1],
                         y = dim_names[2],
-                        color = feat, alpha = feat
+                        color = feat,
+                        alpha = feat
                     ),
-                    show.legend = show_legend, shape = 19, size = point_size
+                    show.legend = show_legend,
+                    shape = 19,
+                    size = point_size
                 )
             } else {
                 pl <- pl + ggplot2::geom_point(
-                    data = data, aes_string2(
+                    data = data,
+                    aes_string2(
                         x = dim_names[1],
                         y = dim_names[2],
                         color = feat
                     ),
-                    show.legend = show_legend, shape = 19,
+                    show.legend = show_legend,
+                    shape = 19,
                     size = point_size,
                     alpha = point_alpha
                 )
@@ -3512,7 +3539,8 @@ dimFeatPlot2D <- function(
         )
 
         # point shape
-        point_shape <- match.arg(point_shape, choices = c("border", "no_border"))
+        point_shape <- match.arg(
+            point_shape, choices = c("border", "no_border"))
 
         # Set feat_type and spat_unit
         spat_unit <- set_default_spat_unit(
@@ -3561,7 +3589,8 @@ dimFeatPlot2D <- function(
             stop("No `feats` selected to plot.", call. = FALSE)
         }
         selected_feats <- feats
-        selected_feats <- selected_feats[selected_feats %in% rownames(expr_values)]
+        selected_feats <- selected_feats[
+            selected_feats %in% rownames(expr_values)]
         if (length(selected_feats) == 0) {
             stop("Selected `feats` not found in expression information",
                 call. = FALSE
@@ -3582,8 +3611,8 @@ dimFeatPlot2D <- function(
                 selected_feats
             )
         } else {
-            subset_expr_data <- expr_values[rownames(expr_values) %in%
-                selected_feats, ]
+            subset_expr_data <- expr_values[
+                rownames(expr_values) %in% selected_feats, ]
             t_sub_expr_data <- t_flex(subset_expr_data)
             t_sub_expr_data_DT <- data.table::as.data.table(
                 as.matrix(t_sub_expr_data)
@@ -3618,13 +3647,15 @@ dimFeatPlot2D <- function(
             feat_type = feat_type
         )
 
-        annotated_DT <- data.table::merge.data.table(cell_metadata,
+        annotated_DT <- data.table::merge.data.table(
+            cell_metadata,
             dim_DT,
             by = "cell_ID"
         )
 
         ## merge feat info
-        annotated_feat_DT <- data.table::merge.data.table(annotated_DT,
+        annotated_feat_DT <- data.table::merge.data.table(
+            annotated_DT,
             t_sub_expr_data_DT,
             by = "cell_ID"
         )
@@ -3653,7 +3684,8 @@ dimFeatPlot2D <- function(
                 by.x = "from", by.y = "cell_ID"
             )
             from_dim_names <- paste0("from_", old_dim_names)
-            data.table::setnames(annotated_network_DT,
+            data.table::setnames(
+                annotated_network_DT,
                 old = old_dim_names,
                 new = from_dim_names
             )
@@ -3663,7 +3695,8 @@ dimFeatPlot2D <- function(
                 by.x = "to", by.y = "cell_ID"
             )
             to_dim_names <- paste0("to_", old_dim_names)
-            data.table::setnames(annotated_network_DT,
+            data.table::setnames(
+                annotated_network_DT,
                 old = old_dim_names,
                 new = to_dim_names
             )
@@ -3728,7 +3761,8 @@ dimFeatPlot2D <- function(
                 cow_n_col = cow_n_col,
                 nr_plots = length(savelist)
             ),
-            rel_heights = cow_rel_h, rel_widths = cow_rel_w,
+            rel_heights = cow_rel_h,
+            rel_widths = cow_rel_w,
             align = cow_align
         )
 
@@ -3743,7 +3777,8 @@ dimFeatPlot2D <- function(
             do.call(
                 "all_plots_save_function",
                 c(list(
-                    gobject = gobject, plot_object = combo_plot,
+                    gobject = gobject,
+                    plot_object = combo_plot,
                     default_save_name = default_save_name
                 ), save_param)
             )
@@ -3973,17 +4008,24 @@ spatDimFeatPlot2D <- function(
         ncol <- 1
         nrow <- 2
         combo_plot <- cowplot::plot_grid(
-            dmpl, spl,
-            ncol = ncol, nrow = nrow, rel_heights = c(1),
-            rel_widths = c(1), align = "v"
+            dmpl,
+            spl,
+            ncol = ncol,
+            nrow = nrow,
+            rel_heights = c(1),
+            rel_widths = c(1),
+            align = "v"
         )
     } else {
         ncol <- 2
         nrow <- 1
         combo_plot <- cowplot::plot_grid(
             dmpl, spl,
-            ncol = ncol, nrow = nrow, rel_heights = c(1),
-            rel_widths = c(1), align = "h"
+            ncol = ncol,
+            nrow = nrow,
+            rel_heights = c(1),
+            rel_widths = c(1),
+            align = "h"
         )
     }
 
@@ -4722,7 +4764,8 @@ spatDimCellPlot2D <- function(
     if (plot_alignment == "vertical") {
         ncol <- 1
         nrow <- 2
-        combo_plot <- cowplot::plot_grid(dmpl, spl,
+        combo_plot <- cowplot::plot_grid(
+            dmpl, spl,
             ncol = ncol, nrow = nrow,
             rel_heights = c(1), rel_widths = c(1),
             align = "v"
@@ -4730,7 +4773,8 @@ spatDimCellPlot2D <- function(
     } else {
         ncol <- 2
         nrow <- 1
-        combo_plot <- cowplot::plot_grid(dmpl, spl,
+        combo_plot <- cowplot::plot_grid(
+            dmpl, spl,
             ncol = ncol, nrow = nrow,
             rel_heights = c(1), rel_widths = c(1),
             align = "h"